refactor: зависимость только от IController
Теперь для подключения библиотеки достаточно подключить только интерфейс
контроллера.
Вероятно, костыль: для сохранения взаимодействия объектов пришлось
наследовать ControllerImpl от std::enable_shared_from_this<T>.
