+++ include/core/ControllerImpl.h
@@ -3,7 +3,7 @@
 #include "interfaces/IController.h"
 #include "interfaces/IModel.h"
 
-class ControllerImpl: public IController
+class ControllerImpl: public IController, public std::enable_shared_from_this<ControllerImpl>
 {
 public:
 	explicit ControllerImpl(IModel::Ptr model);
