+++ src/core/createInstance.cpp
@@ -1,0 +1,11 @@
+#include "interfaces/IController.h"
+#include "interfaces/IModel.h"
+#include "core/Controller.h"
+#include "core/Model.h"
+
+IController::Ptr createInstance()
+{
+	const auto model = createModel();
+	const auto controller = createController(model);
+	return controller;
+}
